+++ Editor/Playa/RendererGroup/SaintsRendererGroup.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using SaintsField.Editor.Linq;
 using SaintsField.Editor.Playa.Utils;
 using SaintsField.Playa;
 using SaintsField.Utils;