+++ Samples~/Scripts/Issue420/Issue420Mono.cs
@@ -1,3 +1,5 @@
+using SaintsField.Playa;
+
 namespace SaintsField.Samples.Scripts.Issue420
 {
     public class Issue420Mono : SaintsMonoBehaviour
@@ -8,8 +10,10 @@ namespace SaintsField.Samples.Scripts.Issue420
         //
         // public FloatWithBaseValue myFloat1;
         // public FloatWithBaseValue myFloat2;
+        [LayoutStart("This is a Label Field layout", ELayout.LabelField)]
+        [AboveText(nameof(myFloat3), paddingLeft: 0)]
+
         [SaintsRow(inline: true)]
-        [LabelText("OK")]
         public FloatWithBaseValue myFloat3;
     }
 }