+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 5.25.0 ##
+
+1.  Add: `Layout` now support `ELayout.LabelField` [#420](https://github.com/TylerTemp/SaintsField/issues/420)
+2.  Fix: `Toggle` in horizontal layout not get current label position
+
 ## 5.24.0 ##
 
 1.  Unity Fix: You can now enable UI Toolkit support for Animator, which allows you to use UI Toolkit attribute/field type in `StateMachineBehavior`. Menu: `Tools` - `SaintsField` - `Enable UI Toolkit To Animator`